requirements.txt
alembic==1.16.1
anyio==4.9.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
python-telegram-bot==22.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
